Removing scene delegate - not supported in iOS 11 which is a requirement of the tech test.
